주소, 일정 컴포넌트 분리
