+++ src/app/(myPage)/createactivity/_actions/createActions.ts
@@ -44,7 +44,8 @@ export async function createActions(formData: FormData): Promise<void> {
   );
 
   if (!response.ok) {
-    console.error(error);
+    const errorData = await response.json().catch(() => ({}));
+    console.error('폼 제출 실패:', response.status, errorData);
   } else {
     console.log('폼 제출 성공');
   }
