+++ src/store/createActivityStore.ts
@@ -1,0 +1,4 @@
+// store/useActivityStore.ts
+import { create } from 'zustand';
+
+export const createActivityStore = create(() => {});
